chore: moved msgs creation methods from builders to enums
