+++ src/relationships/mod.rs
@@ -4,7 +4,6 @@ pub mod mock;
 pub mod models;
 pub mod models_query;
 pub mod msg;
-pub mod msg_builder;
 #[cfg(feature = "query")]
 pub mod querier;
 #[cfg(feature = "query")]
