+++ src/subspaces/mod.rs
@@ -3,7 +3,6 @@ pub mod mock;
 
 pub mod models;
 pub mod msg;
-pub mod msg_builder;
 #[cfg(feature = "query")]
 pub mod querier;
 #[cfg(feature = "query")]
